graph

+++ scripts/metadata.py
@@ -3,13 +3,15 @@ import json
 import os
 import sys
 import shutil
+import time
 
 ALGORITHM = sys.argv[1]
+THRESHOLD = int(sys.argv[2])
 
-NUCLIPATH = 'graphOmahaIndexed.mtx_'+ALGORITHM+'_NUCLEI'
-CIRCLPATH = 'graphOmahaIndexed.mtx_'+ALGORITHM+'_circle.json'
-INDEXPATH = 'indexDict.json'
-ONESPATH = 'Omahagraph.json'
+NUCLIPATH = 'graphIndexed.mtx_'+ALGORITHM+'_NUCLEI'
+CIRCLPATH = 'graphIndexed.mtx_'+ALGORITHM+'_circle.json'
+INDEXPATH = 'data/indexDict.json'
+ONESPATH  = 'graphOmaha.json'
 OUTPATH   = 'data/nodes/'
 
 def assure_path_exists(path):
@@ -36,75 +38,90 @@ def openNucliFile():
 
 def createNodeFiles(index_data, nucli_data):
 
-    assure_path_exists(OUTPATH)
+    #assure_path_exists(OUTPATH)
     with open(ONESPATH, 'r') as index_file:
         json_data = json.load(index_file)
         index_file.close()
 
+    node_ones = {}
     for line in nucli_data:
         line = line.split(' ')
         index = line[0]
         devices = line[7:-1]
 
-        entry = {}
         dups = {}
         original_ids = []
-
-        for device in devices:
-            new = index_data[device]
-            if new not in dups:
-                original_ids.append(new)
-                dups.update({new:0})
-
-        entry[index] = original_ids
-        ones = []
-        others = []
-        for i in range(0, len(devices)-1, 2):
-            check = index_data[devices[i]] + "." + index_data[devices[i+1]]
-            #print(check)
-            if check in json_data:
-                ones.append([index_data[devices[i]],index_data[devices[i+1]]])
-                #print("y")
-            else:
-                others.append([index_data[devices[i]],index_data[devices[i+1]]])
-        entry[1]=ones
-        entry["0 and -1"]=others
-
         file_name = OUTPATH + index + '.json'
-        with open(file_name, 'w') as out_file:
-            json.dump(entry, out_file)
-            out_file.close()
 
+        if len(devices) > THRESHOLD: #and not os.path.isfile(file_name):
+
+            entry = {}
+            for device in devices:
+                new = index_data[device]
+                if new not in dups:
+                    original_ids.append(new)
+                    dups.update({new:0})
+
+            # t0 = time.time()
+            dups = {}
+            entry[index] = original_ids
+            ones = []
+            others = []
+            for i in range(0, len(devices)-1, 2):
+                check = index_data[devices[i]] + "." + index_data[devices[i+1]]
+                if check not in dups:
+                    #print(check)
+                    if check in json_data:
+                        ones.append([index_data[devices[i]],index_data[devices[i+1]]])
+                        if index not in node_ones:
+                            node_ones.update({index:1})
+                        #print("y")
+                    else:
+                        others.append([index_data[devices[i]],index_data[devices[i+1]]])
+                    dups.update({check:0})
+
+            entry[1]=ones
+            entry["0 and -1"]=others
+
+            with open(file_name, 'w') as out_file:
+                json.dump(entry, out_file, indent=4)
+                out_file.close()
+
+    with open('node_ones.json','w') as out_file:
+        json.dump(node_ones, out_file, indent=4)
+        out_file.close()
 
 def main():
 
-
     index_data = openIndexFile()
     nucli_data = openNucliFile()
-    FILEINPUT   = 'Omahagraph.txt'
-    FILEOUTPUT  = 'Omahagraph.json'
+    FILEINPUT   = 'graphOmaha'
+    FILEOUTPUT  = 'graphOmaha.json'
 
-    entry = dict()
-    data = dict()
+    data = {}
 
     with open(FILEINPUT, 'r') as fileinput:
         lines = fileinput.readlines()
         for line in lines:
+            entry = {}
             nodeid = line.split('\t')
-            if nodeid[2] == "1":
+            edgeValue = nodeid[4].split('\n')[0]
+            if edgeValue == "1":
                 conbi = nodeid[0] + "." + nodeid[1]
-                entry[conbi] = nodeid[2]
+                entry[conbi] = nodeid[4]
                 data.update(entry)
                 conbi = nodeid[1] + "." + nodeid[0]
-                entry[conbi] = nodeid[2]
+                entry[conbi] = nodeid[4]
                 data.update(entry)
     fileinput.close()
+    print('\tCreated temporary file \'Omahagraph.json\'')
 
     with open(FILEOUTPUT, 'w') as fileoutput:
         json.dump(data, fileoutput, indent=4)
     fileoutput.close()
 
     createNodeFiles(index_data, nucli_data)
+    #os.remove(FILEOUTPUT)
 
 if __name__ == '__main__':
     main()
